Do not update the map if the selected filter does not contain an imagePoint corresponding to the currently selected imagePoint
